Moved types from index to prevent future lint errors
