+++ libs/content-search-toolkit/src/types/search.ts
@@ -1,4 +1,4 @@
-import { elasticTagField } from './index'
+import { elasticTagField } from './shared'
 
 export interface SearchInput {
   queryString: string
